fix: make auto-configure after sensitive to spring boot 3 and 4 auto configuration classes
