+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/spring/configuration/JsonMapperConfiguration.java
@@ -19,13 +19,19 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.camunda.client.api.JsonMapper;
 import io.camunda.client.impl.CamundaObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
+@AutoConfiguration
+@AutoConfigureAfter(
+    name = {
+      "org.springframework.boot.jackson2.autoconfigure.Jackson2AutoConfiguration",
+      "org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration"
+    })
 public class JsonMapperConfiguration {
-  private final ObjectMapper objectMapper;
+  final ObjectMapper objectMapper;
 
   @Autowired
   public JsonMapperConfiguration(@Autowired(required = false) final ObjectMapper objectMapper) {
